# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sistema de Frequencia Avance",
    "version": "1.1.0",
    "description": "Documentacao oficial da API versionada em /api/v1 com envelope padrao de resposta."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versao atual"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success", "data", "message", "errors"],
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "nullable": true,
            "description": "Payload principal de sucesso."
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string", "nullable": true },
                "location": { "type": "string", "nullable": true },
                "message": { "type": "string" },
                "code": { "type": "string", "nullable": true },
                "value": { "nullable": true }
              }
            }
          }
        }
      },
...
      "ResponsavelCreateRequest": {
        "type": "object",
        "required": ["nomeResponsavel", "email", "telefone", "dataNascimento"],
        "properties": {
          "alunoId": { "type": "integer", "minimum": 1 },
          "nomeAluno": { "type": "string" },
          "nomeResponsavel": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "telefone": { "type": "string" },
          "dataNascimento": { "type": "string", "format": "date" }
        }
      },
...
    },
    "responses": {
      "Api200": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "ok": {
                "value": {
                  "success": true,
                  "data": { "id": 1, "nome": "Joao" },
                  "message": "Operacao realizada com sucesso.",
                  "errors": null
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erro de validacao",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "validation": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Erro de validacao nos dados enviados.",
                  "errors": [
                    {
                      "field": "page",
                      "location": "query",
                      "message": "page deve ser um inteiro maior que 0.",
                      "value": "0"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Nao autenticado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "missing-token": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Token de acesso ausente.",
                  "errors": [
                    { "code": "AUTH_TOKEN_MISSING" }
                  ]
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/responsaveis/alunos": {
      "get": {
        "tags": ["Responsaveis"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Lista alunos para vinculo de responsavel",
        "parameters": [
          { "name": "salaId", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "includeMeta", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["id", "nome"] } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/responsaveis/aluno/{alunoId}": {
      "get": {
        "tags": ["Responsaveis"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Lista responsaveis de um aluno",
        "parameters": [
          { "name": "alunoId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/responsaveis": {
      "post": {
        "tags": ["Responsaveis"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Cadastra responsavel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResponsavelCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "409": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  }
}